# docs/migration/phase1/postgres-openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learnova Backend",
    "description": "Backend API for Learnova learner and instructor workflows.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses",
        "description": "This returns the learner dashboard payload for the authenticated user.",
        "operationId": "list_courses_courses_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/enroll": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Post Course Enrollment",
        "description": "This enrolls the current learner in a free self-enrollable course.",
        "operationId": "post_course_enrollment_courses__course_slug__enroll_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course",
        "description": "This returns the learner-facing detail payload for one course.",
        "operationId": "get_course_courses__course_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/reviews": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Reviews",
        "description": "This returns the review payload for one course.",
        "operationId": "get_course_reviews_courses__course_slug__reviews_get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Post Course Review",
        "description": "This creates or updates the current learner review for the target course.",
        "operationId": "post_course_review_courses__course_slug__reviews_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/payments/order": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Post Payment Order",
        "description": "This creates a Razorpay order for a paid course enrollment.",
        "operationId": "post_payment_order_courses__course_slug__payments_order_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/payments/verify": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Post Payment Verification",
        "description": "This verifies the Razorpay signature and marks the course enrollment as paid.",
        "operationId": "post_payment_verification_courses__course_slug__payments_verify_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/content/{content_slug}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Content",
        "description": "This returns a single content item payload for direct lesson-player loading.",
        "operationId": "get_course_content_courses__course_slug__content__content_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "content_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/content/{content_slug}/progress": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Post Course Content Progress",
        "description": "This updates lesson/document/video progress for the current learner.",
        "operationId": "post_course_content_progress_courses__course_slug__content__content_slug__progress_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "content_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentProgressUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/quizzes/{content_slug}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Quiz",
        "description": "This returns the quiz payload for the requested course content slug.",
        "operationId": "get_quiz_courses__course_slug__quizzes__content_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "content_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_slug}/quizzes/{content_slug}/attempts": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Post Quiz Attempt",
        "description": "This scores and stores a learner quiz attempt.",
        "operationId": "post_quiz_attempt_courses__course_slug__quizzes__content_slug__attempts_post",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "content_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizAttemptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/courses": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Courses",
        "description": "This returns the instructor/admin course list for backoffice views.",
        "operationId": "get_admin_courses_admin_courses_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Post Admin Course",
        "description": "This creates a new course record and attaches its initial tags.",
        "operationId": "post_admin_course_admin_courses_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/courses/{course_slug}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Course Detail",
        "description": "This returns one course record for the instructor course form.",
        "operationId": "get_admin_course_detail_admin_courses__course_slug__get",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Put Admin Course",
        "description": "This updates the editable course fields and course tags.",
        "operationId": "put_admin_course_admin_courses__course_slug__put",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Course",
        "description": "This deletes a course and its dependent rows through database cascades.",
        "operationId": "delete_course_admin_courses__course_slug__delete",
        "parameters": [
          {
            "name": "course_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentProgressUpdateRequest": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(not_started|in_progress|completed)$",
            "title": "Status"
          },
          "lastPosition": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Lastposition",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ContentProgressUpdateRequest"
      },
      "CourseCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "shortDescription": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Shortdescription"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "coverImageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimageurl"
          },
          "websiteId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Websiteid"
          },
          "visibility": {
            "type": "string",
            "pattern": "^(everyone|signed_in)$",
            "title": "Visibility",
            "default": "everyone"
          },
          "accessRule": {
            "type": "string",
            "pattern": "^(open|invitation|payment)$",
            "title": "Accessrule",
            "default": "open"
          },
          "price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Price",
            "default": 0
          },
          "responsibleUserId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsibleuserid"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "isPublished": {
            "type": "boolean",
            "title": "Ispublished",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "shortDescription"
        ],
        "title": "CourseCreateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentVerificationRequest": {
        "properties": {
          "razorpayOrderId": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Razorpayorderid"
          },
          "razorpayPaymentId": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Razorpaypaymentid"
          },
          "razorpaySignature": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Razorpaysignature"
          }
        },
        "type": "object",
        "required": [
          "razorpayOrderId",
          "razorpayPaymentId",
          "razorpaySignature"
        ],
        "title": "PaymentVerificationRequest"
      },
...
      "QuizAttemptRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/QuizAnswerRequest"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "QuizAttemptRequest"
      },
...
      "ReviewSubmissionRequest": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "rating",
          "comment"
        ],
        "title": "ReviewSubmissionRequest"
      },
...
    }
  }
}